답변(RePost)의 필드 정보 변경 및 xss 방지 기능 강화.

+++ src/main/java/com/forweaver/domain/Project.java
@@ -40,7 +40,7 @@ public class Project implements Serializable {
 	private String originalProject;
 	@DBRef
 	private Weaver creator;
-	private int push;
+	private long push;
 	@DBRef
 	private List<Project> childProjects = new ArrayList<Project>();
 	
@@ -171,11 +171,11 @@ public class Project implements Serializable {
 	}
 
 
-	public int getPush() {
+	public long getPush() {
 		return push;
 	}
 
-	public void setPush(int push) {
+	public void setPush(long push) {
 		this.push = push;
 	}
